move IConsolidatorAdapter to types file
